Move style helper functions to library
