+++ src/ui-fragment/src/utils.tsx
@@ -47,6 +47,49 @@ export const setItemInState = (item: any, state: any, setState: (state: any) => 
 	});
 };
 
+export const getAllComponentStyleClasses = (componentObj: any, fragments: any[]) => {
+	let styleClasses: any = {};
+
+	// convert into an object so all classes are unique
+	componentObj.cssClasses?.forEach((cssClass: any) => {
+		// NOTE do we need to merge them deeply?
+		styleClasses[cssClass.id] = cssClass;
+	});
+
+	componentObj.items?.map((co: any) => {
+		const coClasses = getAllComponentStyleClasses(co, fragments);
+		styleClasses = {
+			...styleClasses,
+			...coClasses
+		};
+
+		if (co.type === 'fragment') {
+			const fragment = fragments.find(f => f.id === co.fragmentId);
+
+			styleClasses = {
+				...styleClasses,
+				// we can't avoid this without a messy declare+reassign+export
+				// eslint-disable-next-line @typescript-eslint/no-use-before-define
+				...getAllFragmentStyleClasses(fragment || {}, fragments)
+			};
+		}
+	});
+
+	return styleClasses;
+};
+
+export const getAllFragmentStyleClasses = (fragment: any, fragments: any[] = []) => {
+	if (!fragment || !fragment.data) {
+		return [];
+	}
+
+	const allClasses = {
+		...getAllComponentStyleClasses(fragment, fragments),
+		...getAllComponentStyleClasses(fragment.data, fragments)
+	};
+	return Object.values(allClasses);
+};
+
 export const renderComponents = (state: any, setState: (state: any) => void, setGlobalState: (state: any) => void) => {
 	switch (state.type) {
 		case 'accordion':
